refac after review
