+++ viewer/packages/utilities/index.ts
@@ -19,6 +19,7 @@ export { AutoDisposeGroup } from './src/three/AutoDisposeGroup';
 export { BoundingBoxLOD } from './src/three/BoundingBoxLOD';
 export { toThreeBox3 } from './src/three/toThreeBox3';
 export { unionBoxes } from './src/three/unionBoxes';
+export { determineCurrentDevice, DeviceDescriptor } from './src/device';
 
 export { isBox3OnPositiveSideOfPlane } from './src/three/isBox3OnPositiveSideOfPlane';
 export { visitBox3CornerPoints } from './src/three/visitBox3CornerPoints';
